Add models for user uploading files, tags, and metadata
This adds support for quoting and sharing posts of other user,
tracking metadata like views and impressions of the post.

This also allow support for adding multiple files to a single post.

Similarly, it creates tags from the content of the posts. Also,
the filename of user uploaded files is randomized too.