Add PostManager and basic api for `Post`
This adds a `PostManager` that manages the sharing, quoting,
and toggling like of the post.

Also, `Post` creation using standard apis and getting child/parent
of the post is also implemented. Childs of the post can also be
obtained using `get_childs` whereas parent of the post could be
obtained using get_parent`.

All the media associated with a post can also be retrieve using `get_medias` api.